fast_backend/requirements.txt
aiohttp==3.9.1
aioredis==2.0.1
aiosignal==1.3.1
aiosmtplib==1.1.7
aiosqlite==0.17.0
alembic==1.7.7
annotated-types==0.6.0
anyio==3.6.1
APScheduler==3.10.4
asgiref==3.5.2
async-timeout==4.0.3
asyncpg==0.26.0
attrs==21.4.0
bcrypt==3.2.2
blinker==1.7.0
boto3==1.33.12
botocore==1.33.13
certifi==2022.9.24
cffi==1.15.0
charset-normalizer==2.1.1
ciso8601==2.3.1
click==7.1.2
colorama==0.4.6
coverage==6.5.0
cryptography==37.0.2
dependency-injector==4.39.1
dnspython==2.5.0
ecdsa==0.17.0
email-validator==1.3.1
execnet==1.9.0
fakeredis==2.21.1
fastapi==0.109.2
fastapi-mail==0.4.1
fastapi-migrate==0.1.1
frozenlist==1.4.1
future==0.18.3
geographiclib==2.0
geopy==2.4.1
greenlet==1.1.2
h11==0.13.0
httpcore==0.16.3
httptools==0.4.0
httpx==0.23.3
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
influxdb==5.3.1
influxdb-client==1.30.0
iniconfig==1.1.1
isort==5.10.1
Jinja2==3.1.2
jmespath==1.0.1
loguru==0.6.0
Mako==1.2.0
MarkupSafe==2.1.1
msgpack==1.0.7
multidict==6.0.4
netaddr==0.9.0
netmiko==4.2.0
ntc_templates==4.1.0
numpy==1.26.2
packaging==21.3
pandas==2.1.1
paramiko==3.3.1
passlib==1.7.4
ping3==4.0.4
plogger==1.0.8
pluggy==1.0.0
ply==3.11
provide==0.8
psycopg2-binary==2.9.4
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pycryptodomex==3.19.0
pydantic==1.9.1
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==3.0.9
pyserial==3.5
pysmi==0.3.4
pysnmp==4.4.12
pyspnego==0.10.2
pytest==7.1.2
pytest-asyncio==0.19.0
pytest-cov==4.0.0
pytest-forked==1.4.0
pytest-mock==3.9.0
pytest-xdist==2.5.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
python-nmap==0.7.1
pytz==2022.1
pywinos==2.0.8
pywinrm==0.4.3
PyYAML==6.0
redis==5.0.1
requests==2.28.1
requests-credssp==2.0.0
requests-ntlm==1.2.0
rfc3986==1.5.0
rsa==4.8
Rx==3.2.0
s3transfer==0.8.2
scp==0.14.5
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.36
sqlalchemy2-stubs==0.0.2a27
sqlmodel==0.0.8
starlette==0.36.3
textfsm==1.1.3
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.12
uvicorn==0.20.0
watchfiles==0.17.0
watchgod==0.8.2
websockets==10.3
xmltodict==0.13.0
yarl==1.9.4
zipp==3.8.0
fastapi-mail==0.4.1
geopy==2.4.1